Unit::idの戻り値型を変更
出力などに使うとうまく動作しないため．

+++ include/impl/unit.hpp
@@ -49,7 +49,7 @@ inline const char Unit::name() const
 {
 	return nameList[mID];
 }
-inline const uint8_t Unit::id() const
+inline const int Unit::id() const
 {
 	return mID;
 }
